+++ src/components/AddItem/docs/index.story.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as examples from './examples';
 import {
+  description,
   header,
   api,
   divider,
@@ -13,11 +14,16 @@ import {
   title,
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
-import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
+import {
+  autoSettingsPanel,
+  settingsPanel,
+} from '../../../../stories/utils/SettingsPanel';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import * as AddItemWiringExampleRaw from '!raw-loader!./AddItemWiringExample.tsx';
 import * as AddItemWiringExampleCSSRaw from '!raw-loader!./AddItemWiringExample.st.css';
 import { AddItemWiringExample } from './AddItemWiringExample';
 import { AddItem, ALIGNMENT, ICON_SIZE, DIRECTION } from '../';
+import { storyComponent } from '../../../../stories/helperComponents/storyComponent';
 
 const code = config =>
   baseCode({ components: allComponents, compact: true, ...config });
@@ -25,7 +31,7 @@ const code = config =>
 export default {
   category: 'Components',
   storyName: 'AddItem',
-  component: AddItem,
+  component: storyComponent(AddItem),
   componentPath: '../AddItem.tsx',
   componentProps: () => ({
     'data-hook': 'storybook-AddItem',
@@ -45,6 +51,10 @@ export default {
       tab({
         title: 'Usage',
         sections: [
+          description(
+            'Add Item is a component used to add new items to an existing items list.',
+          ),
+
           importExample({
             source: examples.importExample,
           }),
@@ -64,9 +74,10 @@ export default {
       }),
 
       ...[
+        { title: 'Playground', sections: [playground(), autoSettingsPanel()] },
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
-        { title: 'Playground', sections: [playground()] },
         {
           title: 'Settings Panel',
           sections: [