+++ src/components/Avatar/docs/index.story.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { Avatar } from '../';
 import * as examples from './examples';
 import {
+  description,
   header,
   api,
   divider,
@@ -14,6 +15,7 @@ import {
   title,
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import { AvatarSize } from '../Avatar';
 
 const code = config =>
@@ -40,6 +42,10 @@ export default {
       tab({
         title: 'Usage',
         sections: [
+          description(
+            'Avatar is a type of element that visually represents a user, either as an image, placeholder or text (name initials).',
+          ),
+
           importExample({
             source: examples.importExample,
           }),
@@ -58,9 +64,10 @@ export default {
       }),
 
       ...[
-        { title: 'API', sections: [api()] },
-        { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
+        { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
+        { title: 'TestKit', sections: [testkit()] },
       ].map(tab),
     ]),
   ],