+++ src/components/CalendarPopover/docs/index.story.tsx
@@ -14,52 +14,21 @@ import {
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
 import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import * as CalendarPopoverWiringExampleRaw from '!raw-loader!./CalendarPopoverWiringExample.tsx';
 import * as CalendarPopoverWiringExampleCSSRaw from '!raw-loader!./CalendarPopoverWiringExample.st.css';
 import { CalendarPopoverWiringExample } from './CalendarPopoverWiringExample';
 import { CalendarPopover } from '..';
 import { TPAComponentsProvider } from '../../TPAComponentsConfig';
+import { storyComponent } from '../../../../stories/helperComponents/storyComponent';
 
 const code = config =>
   baseCode({ components: allComponents, compact: true, ...config });
 
-function ExamplePopover(props) {
-  const [rtl, setRtl] = React.useState(false);
-  const rootRef = React.useRef<HTMLDivElement>();
-
-  React.useEffect(() => {
-    if (rootRef && rootRef.current) {
-      const observer = new MutationObserver(mutationsList => {
-        mutationsList.map(mutation => {
-          if (mutation.attributeName === 'dir') {
-            setRtl(
-              (rootRef.current.parentNode as any).getAttribute('dir') === 'rtl',
-            );
-          }
-        });
-      });
-
-      observer.observe(rootRef.current.parentNode, { attributes: true });
-
-      return function cleanup() {
-        observer.disconnect();
-      };
-    }
-  }, [rootRef]);
-
-  return (
-    <div ref={rootRef}>
-      <TPAComponentsProvider value={{ mobile: false, rtl }}>
-        <CalendarPopover {...props} />
-      </TPAComponentsProvider>
-    </div>
-  );
-}
-
 export default {
   category: 'Components',
   storyName: 'CalendarPopover',
-  component: ExamplePopover,
+  component: storyComponent(CalendarPopover),
   componentPath: '../CalendarPopover.tsx',
   componentProps: () => ({
     'data-hook': 'storybook-CalendarPopover',
@@ -97,6 +66,7 @@ export default {
 
       ...[
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
         {