+++ src/components/Checkbox/docs/index.story.tsx
@@ -15,6 +15,7 @@ import {
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
 import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import * as ExtendedRawSource from '!raw-loader!./CheckboxExtendedExample.tsx';
 import * as ExtendedCSSRawSource from '!raw-loader!./CheckboxExtendedExample.st.css';
 import { CheckboxExtendedExample } from './CheckboxExtendedExample';
@@ -88,6 +89,7 @@ export default {
 
       ...[
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
         {