+++ src/components/Event/docs/index.story.tsx
@@ -14,10 +14,12 @@ import {
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
 import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import * as EventWiringExampleRaw from '!raw-loader!./EventWiringExample.tsx';
 import * as EventWiringExampleCSSRaw from '!raw-loader!./EventWiringExample.st.css';
 import { EventWiringExample } from './EventWiringExample';
 import { Event } from '../';
+import { storyComponent } from '../../../../stories/helperComponents/storyComponent';
 
 const code = config =>
   baseCode({ components: allComponents, compact: true, ...config });
@@ -25,7 +27,7 @@ const code = config =>
 export default {
   category: 'Components',
   storyName: 'Event',
-  component: Event,
+  component: storyComponent(Event),
   componentPath: '../Event.tsx',
   componentProps: () => ({
     'data-hook': 'storybook-Event',
@@ -61,6 +63,7 @@ export default {
 
       ...[
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
         {