+++ src/components/IconToggle/docs/index.story.tsx
@@ -12,6 +12,7 @@ import {
   testkit,
   title,
 } from 'wix-storybook-utils/Sections';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import { ReactComponent as StarIcon } from '../../../assets/icons/Star.svg';
 import { ReactComponent as HeartIcon } from '../../../assets/icons/Heart.svg';
 
@@ -65,6 +66,7 @@ export default {
 
       ...[
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
       ].map(tab),