+++ src/components/Pagination/docs/examples.ts
@@ -1,0 +1,3 @@
+export const importExample = `import { Pagination } from 'wix-ui-tpa/Pagination';`;
+
+export const basic = `<Pagination currentPage={1} totalPages={10} />`;