+++ src/components/ProgressBar/docs/index.story.tsx
@@ -14,6 +14,7 @@ import {
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
 import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
 import * as ProgressBarWiringExampleRaw from '!raw-loader!./ProgressBarWiringExample.tsx';
 import * as ProgressBarWiringExampleCSSRaw from '!raw-loader!./ProgressBarWiringExample.st.css';
 import { ProgressBarWiringExample } from './ProgressBarWiringExample';
@@ -64,6 +65,7 @@ export default {
 
       ...[
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
         { title: 'Playground', sections: [playground()] },
         {