+++ src/components/ShareButton/docs/index.story.tsx
@@ -1,23 +1,28 @@
+import * as ShareButtonWiringExampleCSSRaw from '!raw-loader!./ShareButtonWiringExample.st.css';
+import * as ShareButtonWiringExampleRaw from '!raw-loader!./ShareButtonWiringExample.tsx';
 import * as React from 'react';
-import * as examples from './examples';
 import {
-  header,
   api,
+  code as baseCode,
+  description,
   divider,
+  header,
   importExample,
   playground,
   tab,
-  code as baseCode,
   tabs,
   testkit,
   title,
 } from 'wix-storybook-utils/Sections';
-import { allComponents } from '../../../../stories/utils/allComponents';
-import { settingsPanel } from '../../../../stories/utils/SettingsPanel';
-import * as ShareButtonWiringExampleRaw from '!raw-loader!./ShareButtonWiringExample.tsx';
-import * as ShareButtonWiringExampleCSSRaw from '!raw-loader!./ShareButtonWiringExample.st.css';
-import { ShareButtonWiringExample } from './ShareButtonWiringExample';
 import { ShareButton } from '../';
+import { allComponents } from '../../../../stories/utils/allComponents';
+import { settingsApi } from '../../../../stories/utils/SettingsApi';
+import {
+  autoSettingsPanel,
+  settingsPanel,
+} from '../../../../stories/utils/SettingsPanel';
+import * as examples from './examples';
+import { ShareButtonWiringExample } from './ShareButtonWiringExample';
 
 const code = config =>
   baseCode({ components: allComponents, compact: true, ...config });
@@ -51,6 +56,10 @@ export default {
       tab({
         title: 'Usage',
         sections: [
+          description(
+            '`ShareButton` is a component allowing to render a button for sharing data using the [`Navigator.share`](https://developer.mozilla.org/en-US/docs/Web/API/Navigator/share) method.',
+          ),
+
           importExample({
             source: examples.importExample,
           }),
@@ -68,9 +77,10 @@ export default {
       }),
 
       ...[
+        { title: 'Playground', sections: [playground(), autoSettingsPanel()] },
         { title: 'API', sections: [api()] },
+        { title: 'Style API', sections: [settingsApi()] },
         { title: 'TestKit', sections: [testkit()] },
-        { title: 'Playground', sections: [playground()] },
         {
           title: 'Settings Panel',
           sections: [