+++ src/components/Tabs/docs/index.story.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Tabs, SKIN, ALIGNMENT, VARIANT } from '..';
 import * as TabsSource from '!raw-loader!../Tabs.tsx';
 import { Examples } from './examples';
-import { TPAComponentsProvider } from '../../TPAComponentsConfig';
+import { storyComponent } from '../../../../stories/helperComponents/storyComponent';
 
 const items = Array(20)
   .fill('')
@@ -13,43 +13,10 @@ const exampleItems = [
   { label: 'many items', value: items },
 ];
 
-function ExampleTabs(props) {
-  const [rtl, setRtl] = React.useState(false);
-  const rootRef = React.useRef<HTMLDivElement>();
-
-  React.useEffect(() => {
-    if (rootRef && rootRef.current) {
-      const observer = new MutationObserver(mutationsList => {
-        mutationsList.map(mutation => {
-          if (mutation.attributeName === 'dir') {
-            setRtl(
-              (rootRef.current.parentNode as any).getAttribute('dir') === 'rtl',
-            );
-          }
-        });
-      });
-
-      observer.observe(rootRef.current.parentNode, { attributes: true });
-
-      return function cleanup() {
-        observer.disconnect();
-      };
-    }
-  }, [rootRef]);
-
-  return (
-    <div ref={rootRef}>
-      <TPAComponentsProvider value={{ mobile: false, rtl }}>
-        <Tabs {...props} />
-      </TPAComponentsProvider>
-    </div>
-  );
-}
-
 export default {
   category: 'Components',
   storyName: 'Tabs',
-  component: ExampleTabs,
+  component: storyComponent(Tabs),
   source: TabsSource,
   componentPath: '../Tabs.tsx',
   componentProps: setState => ({