+++ src/components/TextArea/docs/index.story.tsx
@@ -11,12 +11,14 @@ import {
   tabs,
   testkit,
   title,
+  description,
 } from 'wix-storybook-utils/Sections';
 import { allComponents } from '../../../../stories/utils/allComponents';
 import { autoSettingsPanel } from '../../../../stories/utils/SettingsPanel';
 import { TextAreaConnected as TextArea } from './connected';
 import { TextAreaTheme } from '../TextAreaEnums';
 import { settingsApi } from '../../../../stories/utils/SettingsApi';
+import { storyComponent } from '../../../../stories/helperComponents/storyComponent';
 
 const code = config =>
   baseCode({ components: allComponents, compact: true, ...config });
@@ -24,7 +26,7 @@ const code = config =>
 export default {
   category: 'Components',
   storyName: 'TextArea',
-  component: TextArea,
+  component: storyComponent(TextArea),
   componentPath: '../TextArea.tsx',
   componentProps: () => ({
     'data-hook': 'storybook-TextArea',
@@ -54,6 +56,10 @@ export default {
       tab({
         title: 'Usage',
         sections: [
+          description(
+            '`TextArea` is a component allowing to render a multi-line custom text.',
+          ),
+
           importExample({
             source: examples.importExample,
           }),