+++ src/components/Tooltip/docs/examples.tsx
@@ -1,0 +1,9 @@
+export const importExample = `import { Tooltip } from 'wix-ui-tpa/Tooltip';`;
+
+export const basic = `
+<Tooltip content="Hello World">
+  <span>
+    Hover me for a tooltip!
+  </span>
+</Tooltip>
+`;